(rpc_err) split into its own file
Matching clnt_stat and auth_stat

Remove unused re_flags, RE_flags, and RPC_ERR_FLAGS_*

Swap clnt_vc_ncreate_svc() from svc.h into the more appropriate clnt.h
(the underlying purpose of this patch).
